Moved methods to articleview + updated css
